Extract some objects creation functions from main to make code usable as external library
